+++ pkg/analysis_server/lib/src/services/correction/dart/add_super_constructor_invocation.dart
@@ -4,9 +4,9 @@
 
 import 'package:analysis_server/src/services/correction/dart/abstract_producer.dart';
 import 'package:analysis_server/src/services/correction/fix.dart';
-import 'package:analysis_server/src/services/correction/util.dart';
 import 'package:analyzer/dart/ast/ast.dart';
 import 'package:analyzer/dart/element/element.dart';
+import 'package:analyzer/dart/element/type.dart';
 import 'package:analyzer_plugin/src/utilities/string_utilities.dart';
 import 'package:analyzer_plugin/utilities/change_builder/change_builder_core.dart';
 import 'package:analyzer_plugin/utilities/fixes/fixes.dart';
@@ -112,12 +112,31 @@ class _AddInvocation extends ResolvedCorrectionProducer {
           if (parameter.isNamed) {
             builder.write('${parameter.name}: ');
           }
-          // default value
+          // A default value to pass as an argument.
           builder.addSimpleLinkedEdit(
-              parameter.name, getDefaultValueCode(parameter.type));
+              parameter.name, parameter.type.defaultArgumentCode);
         }
         builder.write(')');
       });
     });
+  }
+}
+
+extension on DartType {
+  String get defaultArgumentCode {
+    if (isDartCoreBool) {
+      return 'false';
+    }
+    if (isDartCoreInt) {
+      return '0';
+    }
+    if (isDartCoreDouble) {
+      return '0.0';
+    }
+    if (isDartCoreString) {
+      return "''";
+    }
+    // No better guess.
+    return 'null';
   }
 }