+++ runtime/bin/dartdev_isolate.h
@@ -58,7 +58,6 @@ class DartDevIsolate {
       Dart_IsolateGroupCreateCallback create_isolate,
       char** packages_file,
       char** script,
-      bool* sound_null_safety,
       CommandLineOptions* dart_options);
 
  protected:
@@ -69,7 +68,6 @@ class DartDevIsolate {
     void Run(Dart_IsolateGroupCreateCallback create_isolate,
              char** package_config_override_,
              char** script,
-             bool* force_no_sound_null_safety,
              CommandLineOptions* dart_options);
 
     DartDev_Result result() const { return result_; }
@@ -83,7 +81,6 @@ class DartDevIsolate {
     static DartDev_Result result_;
     static char** script_;
     static char** package_config_override_;
-    static bool* force_no_sound_null_safety_;
     static std::unique_ptr<char*[], void (*)(char**)> argv_;
     static intptr_t argc_;