+++ runtime/bin/gen_snapshot.cc
@@ -112,13 +112,16 @@ static const char* const kSnapshotKindNames[] = {
   V(save_debugging_info, debugging_info_filename)                              \
   V(save_obfuscation_map, obfuscation_map_filename)
 
+// We define sound_null_safety as an unused option here just to make sure
+// scripts that were passing in this option do not break.
 #define BOOL_OPTIONS_LIST(V)                                                   \
   V(compile_all, compile_all)                                                  \
   V(help, help)                                                                \
   V(obfuscate, obfuscate)                                                      \
   V(strip, strip)                                                              \
   V(verbose, verbose)                                                          \
-  V(version, version)
+  V(version, version)                                                          \
+  V(sound_null_safety, sound_null_safety)
 
 #define STRING_OPTION_DEFINITION(flag, variable)                               \
   static const char* variable = nullptr;                                       \
@@ -680,9 +683,6 @@ static int CreateIsolateAndSnapshot(const CommandLineOptions& inputs) {
 
   Dart_IsolateFlags isolate_flags;
   Dart_IsolateFlagsInitialize(&isolate_flags);
-  isolate_flags.null_safety =
-      Dart_DetectNullSafety(nullptr, nullptr, nullptr, nullptr, nullptr,
-                            kernel_buffer, kernel_buffer_size);
   if (IsSnapshottingForPrecompilation()) {
     isolate_flags.obfuscate = obfuscate;
   }